Add new properties and relationships to Entity model
Added `IsExclusive` to `Entity` with a default value of `false`
and mapped it to the `is_exclusive` column in `MyDbContext`.
Introduced a nullable navigation property `App` in `Entity`
to represent its relationship with the `App` entity.

Added `ExternalId` to `EntityValue` and mapped it to the
`external_id` column in `MyDbContext`. Introduced a nullable
navigation property `Entity` in `EntityValue` to represent
its relationship with the `Entity` entity.

Updated `MyDbContext` to include mappings for the new
properties and relationships, enhancing the data model.